Refactor: Differentiate between `int` and `float` attribute types
The `generate_attribute.dart` file was updated to differentiate between integer and floating-point number attributes. The old "number" type is deprecated in favor of "float" and "int". Now when the attribute type is "float" or "number", it is treated as a floating-point number and when it is an "int" it is treated as an integer.